multimodal-rag-system/pyproject.toml
[project]
name = "multimodal-rag-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docling>=2.84.0",
    "fastapi>=0.135.3",
    "google-genai>=1.70.0",
    "langchain>=1.2.15",
    "langchain-google-genai>=4.2.1",
    "langchain-postgres>=0.0.17",
    "langgraph>=1.1.6",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.24",
    "uvicorn>=0.44.0",
]
